+++ source/storage/storage.c
@@ -28,7 +28,7 @@ bool device_storage_t::is_device(const device_t* device) const
 	return true;
 }
 
-int device_storage_t::read(uint64_t lba, int offset, size_t size, void* buffer) const
+int device_storage_t::read(uint64_t lba, size_t offset, size_t size, void* buffer) const
 {
 	if(!buffer || size == (size_t)0)
 		return ERR_INVALID_PARAMETER;
@@ -46,37 +46,37 @@ int device_storage_t::read(uint64_t lba, int offset, size_t size, void* buffer) 
 	 * If doing an unaligned (by sector size) read, or a read thats less then the size of a sector,
 	 * read the first sector into a temporary buffer, and do an offseted copy into <destination>.
 	 */
-	if(offset != 0 || bytes_left < (size_t)m_sector_size - (size_t)offset)
+	if(offset != 0 || bytes_left < m_sector_size - offset)
 	{
-		sector_buffer = (uint8_t*)malloc((size_t)m_sector_size);
+		sector_buffer = (uint8_t*)malloc(m_sector_size);
 		if(!sector_buffer)
 			return ERR_OUT_OF_MEMORY;
 		
-		status = read_sectors(current_lba, 1, sector_buffer);
+		status = read_sectors(current_lba, (size_t)1, sector_buffer);
 		if(status != SUCCESS)
 			goto cleanup;
 	
 		size_t read_size;
-		if(bytes_left < (size_t)m_sector_size - (size_t)offset)
+		if(bytes_left < m_sector_size - offset)
 		{
 			read_size = bytes_left;
 		}
 		else
 		{
-			read_size = (size_t)m_sector_size - (size_t)offset;
+			read_size = m_sector_size - offset;
 			++current_lba;
 		}
 		
-		memcpy(destination, sector_buffer + (uint64_t)offset, read_size);
-		destination += (uint64_t)read_size;
-		bytes_left -= (size_t)read_size;
+		memcpy(destination, sector_buffer + offset, read_size);
+		destination += read_size;
+		bytes_left -= read_size;
 	}
 
 	/* Check if we can read full sectors straight into <buffer> (<destination>) */
-	if(bytes_left >= (size_t)m_sector_size)
+	if(bytes_left >= m_sector_size)
 	{
-		int sectors = bytes_left / m_sector_size;
-		size_t read_size = (uint64_t)sectors * (uint64_t)m_sector_size;
+		size_t sectors = bytes_left / m_sector_size;
+		size_t read_size = sectors * m_sector_size;
 
 		status = read_sectors(current_lba, sectors, destination);
 		if(status != SUCCESS)
@@ -84,7 +84,7 @@ int device_storage_t::read(uint64_t lba, int offset, size_t size, void* buffer) 
 
 		destination += read_size;
 		bytes_left -= read_size;
-		current_lba += (uint64_t)sectors;
+		current_lba += sectors;
 	}
 
 	/* 
@@ -95,7 +95,7 @@ int device_storage_t::read(uint64_t lba, int offset, size_t size, void* buffer) 
 	{
 		if(!sector_buffer)
 		{
-			sector_buffer = (uint8_t*)malloc((size_t)m_sector_size);
+			sector_buffer = (uint8_t*)malloc(m_sector_size);
 			if(!sector_buffer)
 				return ERR_OUT_OF_MEMORY;
 		}
@@ -115,12 +115,12 @@ cleanup:
 	return status;
 }
 
-int device_storage_t::write(uint64_t lba, int offset, size_t size, const void* buffer) const
+int device_storage_t::write(uint64_t lba, size_t offset, size_t size, const void* buffer) const
 {
 	if(!buffer || size == (size_t)0)
 		return ERR_INVALID_PARAMETER;
 	
-	lba += (size_t)offset / (size_t)m_sector_size;		/* If offset is greater than the size of a sector, update the LBA */
+	lba += offset / m_sector_size;							/* If offset is greater than the size of a sector, update the LBA */
 	offset -= (offset / m_sector_size) * m_sector_size;
 
 	int status;
@@ -133,25 +133,25 @@ int device_storage_t::write(uint64_t lba, int offset, size_t size, const void* b
 	 * If writing unaligned (by sector size), or write of size less then the size of a sector,
 	 * read the first sector (<lba>), update the needed bytes on <offset>, and write the sector back to disk.
 	 */
-	if(offset != 0 || size < (size_t)m_sector_size - (size_t)offset)
+	if(offset != 0 || size < m_sector_size - offset)
 	{
-		sector_buffer = (uint8_t*)malloc((size_t)m_sector_size);
+		sector_buffer = (uint8_t*)malloc(m_sector_size);
 		if(!sector_buffer)
 			return ERR_OUT_OF_MEMORY;
 		
-		status = read_sectors(current_lba, 1, sector_buffer);
+		status = read_sectors(current_lba, (size_t)1, sector_buffer);
 		if(status != SUCCESS)
 			goto cleanup;
 		
 		size_t write_size;
-		if(bytes_left < (size_t)m_sector_size - (size_t)offset)
+		if(bytes_left < m_sector_size - offset)
 			write_size = bytes_left;
 		else
-			write_size = (size_t)m_sector_size - (size_t)offset;
+			write_size = m_sector_size - offset;
 
-		memcpy(sector_buffer + (uint64_t)offset, source, write_size);
+		memcpy(sector_buffer + offset, source, write_size);
 
-		status = write_sectors(current_lba++, 1, sector_buffer);
+		status = write_sectors(current_lba++, (size_t)1, sector_buffer);
 		if(status != SUCCESS)
 			goto cleanup;
 		
@@ -160,16 +160,16 @@ int device_storage_t::write(uint64_t lba, int offset, size_t size, const void* b
 	}
 
 	/* Check if we can write full sectors from <source> into <current_lba> */
-	if(bytes_left >= (size_t)m_sector_size)
+	if(bytes_left >= m_sector_size)
 	{
-		int sectors = bytes_left / (size_t)m_sector_size;
-		size_t write_size = sectors * (size_t)m_sector_size;
+		int sectors = bytes_left / m_sector_size;
+		size_t write_size = sectors * m_sector_size;
 		status = write_sectors(current_lba, sectors, source);
 		if(status != SUCCESS)
 			goto cleanup;
 		
 		source += write_size;
-		current_lba += (uint64_t)sectors;
+		current_lba += sectors;
 		bytes_left -= write_size;
 	}
 
@@ -181,18 +181,18 @@ int device_storage_t::write(uint64_t lba, int offset, size_t size, const void* b
 	{
 		if(!sector_buffer)
 		{
-			sector_buffer = (uint8_t*)malloc((size_t)m_sector_size);
+			sector_buffer = (uint8_t*)malloc(m_sector_size);
 			if(!sector_buffer)
 				return ERR_OUT_OF_MEMORY;
 		}
 
-		status = read_sectors(current_lba, 1, sector_buffer);
+		status = read_sectors(current_lba, (size_t)1, sector_buffer);
 		if(status != SUCCESS)
 			goto cleanup;
 		
 		memcpy(sector_buffer, source, bytes_left);
 		
-		status = write_sectors(current_lba, 1, sector_buffer);
+		status = write_sectors(current_lba, (size_t)1, sector_buffer);
 		if(status != SUCCESS)
 			goto cleanup;
 	}
